requirements.txt
# Django Core
Django==5.0.7
asgiref==3.8.1
sqlparse==0.5.1
tzdata==2024.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djangorestframework-api-key==3.0.0
django-cors-headers==4.3.1
django-filter==24.3

# API Documentation
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1

# Environment Management
django-environ==0.11.2
python-decouple==3.8

# Celery & Task Queue
celery==5.4.0
redis==5.0.8
django-celery-beat==2.6.0
amqp==5.2.0
billiard==4.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
kombu==5.3.7
prompt_toolkit==3.0.47
vine==5.1.0

# Utilities
python-dateutil==2.9.0.post0
six==1.16.0
wcwidth==0.2.13
colorama==0.4.6
configparser==7.1.0
PyJWT==2.9.0
cryptography==42.0.8

# Production Server
gunicorn==22.0.0
daphne==4.1.2

# Channels (WebSocket support)
channels==4.1.0
channels-redis==4.2.0

# Image Processing (for user avatars)
Pillow==10.4.0

# Utilities
pytz==2024.1
